visualization_service/requirements.txt
redis
opencv-python
pyyaml
numpy
supervision
streamlit
flask